+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] 2025-11-19
+## [0.7.0] 2025-11-20
 
 In this release, we refocus the library: `typhoon/type` now contains only the types necessary for building
 runtime tools like `is()` or `HMap`. Complex types requiring reflection and/or additional context have been removed